Remove default key generation and associated expectations.
Currently, the npm install process attempts to run install.js via node (or
nodejs if you're on debian) to generate a default set of keys. Rather than
try to make this work for all possible platform/shell combinations, remove
the install script and the logic that attempts to load default keys. Throw
if no keys are specified at initialization.

This is a breaking change for those relying on the old behaviour, but if
you're relying on keying material that lives under node_modules, it's time
to reevaluate your security processes anyway.

Note that tests are now executable via ```npm test```... on every platform
except debian.  (Sorry about that.  ln -s is your friend.)